Updated changed projectBackupId type from Integer to Long
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@2563 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/dao/ProjectBackupDAO.java
@@ -33,7 +33,7 @@ public class ProjectBackupDAO extends GenericDAO<ProjectBackup, Long> {
 
             for (Object o : results) {
                 Object[] projectBackup = (Object[]) o;
-                Integer projectBackupId = (Integer) projectBackup[0];
+                Long projectBackupId = (Long) projectBackup[0];
                 Long projectId = (Long) projectBackup[1];
                 Date backupTime = (Date) projectBackup[2];
                 String backupPath = (String) projectBackup[3];
